[FIXED JENKINS-13751] Allow reference to base SCM in promotion process
Previously, plugins such as the Git Publisher would not work in promotion processes because the SCM reference would be
empty.

This just makes a change to get the base job's SCM when getSCM is called on the promotion process, allowing the use of
steps such as the Git Publisher in promotion processes.