Changed to use WorkspaceList.allocate instead of acquire
Acquiring a quick lock causes parallelised builds to back up waiting for
the workspace to become available.  This is because WorkspaceList.allocate
will give executing jobs the first workspace as the given lock is a quick
lock.  Then builds will wait in WorkspaceList.acquire all trying to get
the same workspace.