Add support for build wrappers in promotions
This commit adds support for build wrappers in the promotion processes.
The build wrappers can be configured in a new "Promotion Environment"
section of the promotion process configuration.

This is useful e.g. in order to be able to prepare an environment or inject
secrets for the promotion.

There are a number of issues calling for this support:
[FIXED JENKINS-23977]
[FIXED JENKINS-25526]
[FIXED JENKINS-43656]

For [JENKINS-14169] this change might provide a workaround by allowing
to inject environment vars for the promotion either from the
configuration or by injecting them from a file passed to the promotion process.