Fix incompatability with Jenkins 2.266 and higher in test code.
both JobDslManualConditionConverterTest and JobDslPromotionProcessConverterTest would fail when run against a newer Jenkins core (for example via the PCT).  The cause was that annotations are no longer automatically processed when serialising / deserialising code.
Fixed this so that we regestier the classes wit interesting annotations.

Also updated the test to use hamcreset to get better errors than binary
pass/fail